refactor(bookmarks): 💡 update service & modal fn
