+++ server/packages/bookmarks/src/service/tag.rs
@@ -1,6 +1,11 @@
+/*
+ * @Author: suxiaoshao suxiaoshao@gmail.com
+ * @Date: 2024-01-06 01:30:13
+ * @LastEditors: suxiaoshao suxiaoshao@gmail.com
+ * @LastEditTime: 2024-03-27 05:32:19
+ */
 use async_graphql::SimpleObject;
 use diesel::PgConnection;
-use std::collections::HashSet;
 use time::OffsetDateTime;
 use tracing::{event, Level};
 
@@ -64,61 +69,10 @@ impl Tag {
         let tags = TagModel::get_by_ids(ids, conn)?;
         Ok(tags.into_iter().map(|x| x.into()).collect())
     }
-    /// 判断标签是否全部存在
-    pub fn exists_all<'a, T: Iterator<Item = &'a i64>>(
-        tags: T,
-        conn: &mut PgConnection,
-    ) -> GraphqlResult<()> {
-        let tag_ids: HashSet<i64> = tags.cloned().collect();
-        let database_tags = TagModel::get_list(conn)?;
-        let database_tags: HashSet<i64> = database_tags.into_iter().map(|tag| tag.id).collect();
-        for id in tag_ids {
-            if !database_tags.contains(&id) {
-                event!(Level::ERROR, "标签不存在: {}", id);
-                return Err(GraphqlError::NotFound("标签", id));
-            }
-        }
-        Ok(())
-    }
 }
 
 /// collection_id 相关
 impl Tag {
-    /// 根据 collection_id 删除标签
-    pub fn delete_by_collection(collection_id: i64, conn: &mut PgConnection) -> GraphqlResult<()> {
-        TagModel::delete_by_collection(collection_id, conn)?;
-        Ok(())
-    }
-    /// 验证 tags 属于 collection_id
-    pub fn belong_to_collection<'a, T: Iterator<Item = &'a i64>>(
-        collection_id: Option<i64>,
-        tags: T,
-        conn: &mut PgConnection,
-    ) -> GraphqlResult<()> {
-        let allow_tags = match collection_id {
-            Some(id) => TagModel::allow_tags(id, conn)?,
-            None => TagModel::query_root(conn)?,
-        };
-        let allow_tags: HashSet<_> = allow_tags.into_iter().map(|tag| tag.id).collect();
-        // 存在不符合的 tags
-        for tag in tags {
-            if !allow_tags.contains(tag) {
-                event!(
-                    Level::ERROR,
-                    "标签: {} 不属于集合: {:?}",
-                    tag,
-                    collection_id
-                );
-                return Err(GraphqlError::Scope {
-                    sub_tag: "标签",
-                    sub_value: *tag,
-                    super_tag: "集合",
-                    super_value: collection_id,
-                });
-            }
-        }
-        Ok(())
-    }
     /// 获取标签列表
     pub fn query(
         collection_id: Option<i64>,
